refactor: ApiResponse를 remote 패키지 하위로 이동
