+++ app/src/main/java/com/gurumlab/vocaroutine/data/source/remote/ApiResponse.kt
@@ -1,4 +1,4 @@
-package com.gurumlab.vocaroutine.data.model
+package com.gurumlab.vocaroutine.data.source.remote
 
 sealed interface ApiResponse<T : Any>
