+++ app/src/main/java/com/gurumlab/vocaroutine/di/GptApiClientModule.kt
@@ -1,6 +1,7 @@
 package com.gurumlab.vocaroutine.di
 
 import com.gurumlab.vocaroutine.BuildConfig
+import com.gurumlab.vocaroutine.data.source.remote.ApiCallAdapterFactory
 import com.gurumlab.vocaroutine.data.source.remote.GptApiClient
 import com.squareup.moshi.Moshi
 import dagger.Module
@@ -53,6 +54,7 @@ object GptModule{
             .baseUrl("https://api.openai.com/")
             .client(client)
             .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .addCallAdapterFactory(ApiCallAdapterFactory.create())
             .build()
     }
